mock ClientResponse _loop attr to avoid crashes
A real ClientResponse object will have ._post_init() called on it to
supply the event loop.  The __del__ method expects an event loop to be
there and will crash if not found.  Mock it to avoid crashes when an
exception outside of aiohttpretty forces shutdown of the ClientResponse
object.

This fixes a bug in the WaterButler-python 3.5 conversion that was
causing asyncs in with pytest.raises to crash.